Moved VisitsTracker service to Visit namespace
